src/requirements.txt
pyyaml
boto3==1.25.5
boto3-stubs==1.25.5
boto3-stubs-lite==1.25.5
botocore==1.28.5
botocore-stubs==1.28.5
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
fastcore==1.5.27
ghapi==1.0.3
idna==3.4
jmespath==1.0.1
mypy==0.982
mypy-boto3-cloudformation==1.25.4
mypy-boto3-dynamodb==1.25.0
mypy-boto3-ec2==1.25.5
mypy-boto3-lambda==1.25.0
mypy-boto3-rds==1.25.1
mypy-boto3-s3==1.25.0
mypy-boto3-sqs==1.25.0
mypy-boto3-ssm==1.25.0
mypy-extensions==0.4.3
packaging==21.3
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
types-awscrt==0.15.1
types-s3transfer==0.6.0.post4
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
